refactor: rename command module to interaction
