+++ internal/interaction/caroline/bye.go
