+++ internal/interaction/caroline/move.go
@@ -115,11 +115,6 @@ func moveCommand(srv *server.Server) func(*discordgo.Session, *discordgo.Interac
 			log.Printf("%s: %s: %s\n", i.Type, util.InteractionName(i), err)
 			return
 		}
-		err = srv.UC.Player.UpdateNPMessage(s, p, q, true)
-		if err != nil {
-			log.Printf("%s: %s: %s\n", i.Type, util.InteractionName(i), err)
-			return
-		}
 
 		err = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
 			Type: discordgo.InteractionResponseChannelMessageWithSource,
