+++ internal/interaction/caroline/np_component.go
