+++ internal/interaction/caroline/remove.go
@@ -118,5 +118,11 @@ func removeCommand(srv *server.Server) func(*discordgo.Session, *discordgo.Inter
 		if err != nil {
 			log.Printf("%s: %s: %s\n", i.Type, util.InteractionName(i), err)
 		}
+
+		err = srv.UC.Player.UpdateNPMessage(s, p, q, true)
+		if err != nil {
+			log.Printf("%s: %s: %s\n", i.Type, util.InteractionName(i), err)
+			return
+		}
 	}
 }
