+++ internal/interaction/caroline/reset.go
