+++ internal/interaction/caroline/stat.go
